+++ pages/index.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import Image from "next/image";
 import Link from "next/link";
 import Head from "next/head";
-import Apy from "../src/components/Apy";
 import Allocation from "../src/components/Allocation";
 import formatSeo from "../src/utils/seo";
 import transformLinks from "../src/utils/transformLinks";
@@ -10,7 +9,7 @@ import Collateral from "../src/components/Collateral";
 import Footer from "../src/components/Footer";
 import Seo from "../src/components/strapi/seo";
 import capitalize from "lodash/capitalize";
-import { Hero, SecretSauce, Faq, Ogv } from "../src/homepage/sections";
+import { Hero, Apy, SecretSauce, Faq, Ogv } from "../src/homepage/sections";
 import { useRouter } from "next/router";
 import { fetchAPI } from "../lib/api";
 import { fetchApy } from "../lib/apy";
