+++ src/homepage/components/index.ts
@@ -1,2 +1,3 @@
 export { default as Animation } from "./Animation";
 export { default as GrowingWallet } from "./GrowingWallet";
+export { default as LineChart } from "./LineChart";
