+++ src/homepage/sections/index.ts
@@ -1,4 +1,5 @@
 export { default as Hero } from "./Hero";
+export { default as Apy } from "./Apy";
 export { default as SecretSauce } from "./SecretSauce";
 export { default as Ogv } from "./Ogv";
 export { default as Faq } from "./Faq";
